Adaptive-SQL-Optimizer/requirements.txt
pandas
scikit-learn
streamlit
mysql-connector-python